pyproject.toml
[tool.poetry]
name = "av-crypto-trading"
version = "0.1.0"
description = "Algorithmic trading using Binance API."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "av_crypto_trading"}]

[tool.poetry.scripts]
trade = "av_crypto_trading.core.cli:main"

[tool.poetry.dependencies]
python = ">3.9.7,<3.11"
pandas = ">=0.23.4"
aiohttp = "3.8.0"
aiosignal = "1.2.0"
appdirs = "1.4.4"
async-timeout = "4.0.1"
attrs = "21.2.0"
backports-zoneinfo = "0.2.1"
black = "21.10b0"
cattrs = "1.10.0"
certifi = "2021.10.8"
charset-normalizer = "2.0.7"
click = "8.0.3"
clikit = "0.6.2"
cmdstanpy = ">=0.9.5"
convertdate = ">=2.1.2"
crashtest = "0.3.1"
cycler = "0.11.0"
cython = ">=0.22"
dateparser = "1.1.0"
ephem = "4.1.3"
frozenlist = "1.2.0"
future = "0.18.2"
gevent = "21.12.0"
greenlet = "1.1.2"
hijri-converter = "2.2.2"
holidays = "0.24"
html5lib = "1.1"
httpstan = "4.6.1"
idna = "3.3"
iniconfig = "1.1.1"
kiwisolver = "1.3.2"
korean-lunar-calendar = "0.2.1"
lunarcalendar = ">=0.0.9"
lxml = ">=4.7.1"
marshmallow = "3.14.1"
matplotlib = ">=2.0.0"
multidict = "5.2.0"
mypy-extensions = "0.4.3"
numpy = "1.26.4"
packaging = ">=21.3"
pandas-datareader = "0.10.0"
pastel = "0.2.1"
pathspec = "0.9.0"
pillow = "8.4.0"
platformdirs = "2.4.0"
plotly = "5.3.1"
pluggy = "1.0.0"
py = "1.11.0"
pydantic = "1.10.8"
pylev = "1.4.0"
pymeeus = "0.5.11"
pyparsing = "2.4.7"
pysimdjson = "3.2.0"
pytest = "6.2.5"
python-binance = "^1.0.15"
python-dateutil = ">=2.8.0"
python-dotenv = "1.0.0"
pytz = ">=2021.3"
pytz-deprecation-shim = ">=0.1.0.post0"
regex = "2021.11.10"
requests = ">=2.26.0"
requests-cache = "0.9.1"
setuptools-git = ">=1.2"
six = "1.16.0"
sqlalchemy = ">=1.4.27"
ta = "0.7.0"
tenacity = "8.0.1"
toml = "0.10.2"
tomli = "1.2.2"
tqdm = ">=4.36.1"
typing-extensions = ">=3.10.0.2"
tzdata = "2022.7" #2021.5
tzlocal = "4.1"
ujson = "4.2.0"
url-normalize = "1.4.3"
urllib3 = "1.26.7"
webargs = "8.1.0"
webencodings = "0.5.1"
websockets = ">=9.1"
yarl = "1.7.2"
zope-event = "4.5.0"
zope-interface = "5.4.0"
yfinance = "^0.2.18"
unicorn-binance-websocket-api = "^1.46.1"
tweepy = "^4.14.0"
nltk = "^3.8.1"
seaborn = "^0.12.2"
spacy = "^3.5.2"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
databases = "^0.7.0"
aiosqlite = ">=0.19.0"
streamlit = "^1.22.0"
pystan = ">=2.14"
prophet = "1.1.2"
psycopg2-binary = "^2.9.6"
pytest-xdist = "^3.3.1"
isort = "^5.12.0"
prospector = "^1.10.2"
pytest-mock = "^3.10.0"
coverage = "^7.2.7"
fastapi-camelcase = "^1.0.5" # maybe can be removed
streamlit-authenticator = "^0.2.2"
streamlit-extras = "^0.2.7"
streamlit-lottie = "^0.0.3" # maybe can be removed
dataclasses-json = "^0.5.7"
en_core_web_sm= {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz"}
tensorflow = "2.15.0"
stable-baselines3 = "2.2.1"
gym-anytrading = "2.0.0"
gym = "0.25.1"
gymnasium = "0.29.1"
finta = "1.3"
mlflow = "2.10.2"
jupyter = "1.0.0"
di-engine = "^0.5.1"
jinja2 = "3.0.3"
itsdangerous = "2.0.1"
werkzeug = "=2.0.3"
shimmy = "0.2.1"
transformers = "^4.38.1"
tensorboard = "2.15"
smartmoneyconcepts = "0.0.24"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


